+++ source/extensions/omni.flux.stage_manager.plugin.filter.usd/omni/flux/stage_manager/plugin/filter/usd/extension.py
@@ -20,12 +20,14 @@ import carb.settings
 import omni.ext
 from omni.flux.stage_manager.factory import get_instance as _get_factory_instance
 
-from .search import SearchPlugin as _SearchPlugin
+from .ignore_prims import IgnorePrimsFilterPlugin as _IgnorePrimsFilterPlugin
+from .omni_prims import OmniPrimsFilterPlugin as _OmniPrimsFilterPlugin
+from .search import SearchFilterPlugin as _SearchFilterPlugin
 
 
 class StageManagerUSDFilterPluginsExtension(omni.ext.IExt):
 
-    _PLUGINS = [_SearchPlugin]
+    _PLUGINS = [_IgnorePrimsFilterPlugin, _OmniPrimsFilterPlugin, _SearchFilterPlugin]
 
     def on_startup(self, _):
         carb.log_info("[omni.flux.stage_manager.plugin.filter.usd] Startup")