+++ source/extensions/omni.flux.stage_manager.plugin.interaction.usd/omni/flux/stage_manager/plugin/interaction/usd/extension.py
@@ -20,12 +20,12 @@ import carb.settings
 import omni.ext
 from omni.flux.stage_manager.factory import get_instance as _get_factory_instance
 
-from .all_prims import AllPrimsPlugin as _AllPrimsPlugin
+from .all_prims import AllPrimsInteractionPlugin as _AllPrimsInteractionPlugin
 
 
 class StageManagerUSDInteractionPluginsExtension(omni.ext.IExt):
 
-    _PLUGINS = [_AllPrimsPlugin]
+    _PLUGINS = [_AllPrimsInteractionPlugin]
 
     def on_startup(self, _):
         carb.log_info("[omni.flux.stage_manager.plugin.interaction.usd] Startup")