+++ source/extensions/omni.flux.stage_manager.plugin.tree.usd/omni/flux/stage_manager/plugin/tree/usd/extension.py
@@ -20,12 +20,13 @@ import carb.settings
 import omni.ext
 from omni.flux.stage_manager.factory import get_instance as _get_factory_instance
 
-from .virtual_groups import VirtualGroupsPlugin as _VirtualGroupsPlugin
+from .prim_groups import PrimGroupsTreePlugin as _PrimGroupsTreePlugin
+from .virtual_groups import VirtualGroupsTreePlugin as _VirtualGroupsTreePlugin
 
 
 class StageManagerUSDTreePluginsExtension(omni.ext.IExt):
 
-    _PLUGINS = [_VirtualGroupsPlugin]
+    _PLUGINS = [_PrimGroupsTreePlugin, _VirtualGroupsTreePlugin]
 
     def on_startup(self, _):
         carb.log_info("[omni.flux.stage_manager.plugin.tree.usd] Startup")