+++ source/extensions/omni.flux.stage_manager.plugin.widget.usd/omni/flux/stage_manager/plugin/widget/usd/base/__init__.py
@@ -15,4 +15,5 @@
 * limitations under the License.
 """
 
+from .state_base import StageManagerStateWidgetPlugin
 from .usd_base import StageManagerUSDWidgetPlugin