+++ source/extensions/lightspeed.trex.app.style/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 ﻿# Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.3.0]
+### Added
+- Added TransparentBackground style
+
 ## [1.2.3]
 ### Added
 - Added styling for Categories Icon