+++ source/extensions/omni.flux.stage_manager.factory/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.1.0]
+### Changed
+- Implemented plugin base classes
+
 ## [1.0.0]
 ### Added
 - Created