+++ source/extensions/omni.flux.stage_manager.plugin.context.usd/docs/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.1.0]
+### Added
+- Added `UsdFileContextPlugin` plugin
+
+### Changed
+- Implemented `CurrentStageContextPlugin` plugin
+
 ## [1.0.0]
 ### Added
 - Created