+++ source/extensions/omni.flux.stage_manager.plugin.filter.usd/docs/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.1.0]
+### Added
+- Added `IgnorePrimsFilterPlugin` plugin
+- Added `OmniPrimsFilterPlugin` plugin
+
 ## [1.0.0]
 ### Added
 - Created